// contracts/generated/typescript/models/MergeConflict.ts
export interface MergeConflict {
    /**
     * 
     * @type {string}
     * @memberof MergeConflict
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof MergeConflict
     */
    proposalId: string;
    /**
     * 
     * @type {string}
     * @memberof MergeConflict
     */
    pageId?: string;
    /**
     * 
     * @type {MergeConflictConflictTypeEnum}
     * @memberof MergeConflict
     */
    conflictType: MergeConflictConflictTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof MergeConflict
     */
    targetBlockId?: string;
    /**
     * 
     * @type {string}
     * @memberof MergeConflict
     */
    targetClaimId?: string;
    /**
     * 
     * @type {string}
     * @memberof MergeConflict
     */
    baseRevisionId?: string;
    /**
     * 
     * @type {string}
     * @memberof MergeConflict
     */
    currentRevisionId?: string;
    /**
     * 
     * @type {{ [key: string]: any; }}
     * @memberof MergeConflict
     */
    baseValue?: { [key: string]: any; };
    /**
     * 
     * @type {{ [key: string]: any; }}
     * @memberof MergeConflict
     */
    currentValue?: { [key: string]: any; };
    /**
     * 
     * @type {{ [key: string]: any; }}
     * @memberof MergeConflict
     */
    proposedValue?: { [key: string]: any; };
    /**
     * 
     * @type {MergeConflictStatusEnum}
     * @memberof MergeConflict
     */
    status: MergeConflictStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof MergeConflict
     */
    resolvedBy?: string;
    /**
     * 
     * @type {{ [key: string]: any; }}
     * @memberof MergeConflict
     */
    resolution?: { [key: string]: any; };
    /**
     * 
     * @type {Date}
     * @memberof MergeConflict
     */
    resolvedAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof MergeConflict
     */
    createdAt: Date;
}


/**
 * @export
 */
export const MergeConflictConflictTypeEnum = {
    Revision: 'revision',
    BlockHash: 'block_hash',
    ClaimState: 'claim_state',
    Semantic: 'semantic'
} as const;
export type MergeConflictConflictTypeEnum = typeof MergeConflictConflictTypeEnum[keyof typeof MergeConflictConflictTypeEnum];

/**
 * @export
 */
export const MergeConflictStatusEnum = {
    Open: 'open',
    Resolved: 'resolved',
    Dismissed: 'dismissed'
} as const;
export type MergeConflictStatusEnum = typeof MergeConflictStatusEnum[keyof typeof MergeConflictStatusEnum];